replace `GaussianImputerBase` with `GaussianImputer`
Since the `GaussianImputer` wasn't doing anything more than the base
class, I just moved all the functionality from the base to the gaussian
imputer and deleted the base.
